+++ src/const/metadata_fields.ts
@@ -16,6 +16,8 @@ export const METADATA_FIELDS_LIST = [
 	"BC-list-note-field",
 	"BC-list-note-neighbour-field",
 	"BC-list-note-exclude-index",
+	"BC-list-note-section",
+	"BC-list-note-exclude",
 	//
 	"BC-dendron-note-field",
 	//
@@ -77,6 +79,12 @@ export const METADATA_FIELDS_MAP: Record<
 	"BC-list-note-exclude-index": {
 		property_type: "checkbox",
 	},
+	"BC-list-note-section": {
+		property_type: "text",
+	},
+	"BC-list-note-exclude": {
+		property_type: "multitext",
+	},
 	//
 	"BC-dendron-note-field": {
 		property_type: "text",
@@ -121,6 +129,8 @@ export const META_ALIAS = {
 	"list-note-field": "BC-list-note-field",
 	"list-note-neighbour-field": "BC-list-note-neighbour-field",
 	"list-note-exclude-index": "BC-list-note-exclude-index",
+	"list-note-section": "BC-list-note-section",
+	"list-note-exclude": "BC-list-note-exclude",
 	//
 	"dendron-note-field": "BC-dendron-note-field",
 	//