+++ src/const/settings.ts
@@ -91,6 +91,8 @@ export const DEFAULT_SETTINGS: BreadcrumbsSettings = {
 
 	self_is_sibling: ["same"],
 
+	exclude_folders: [],
+
 	explicit_edge_sources: {
 		typed_link: {},
 		list_note: {